+++ events/ready.js
@@ -1,4 +1,5 @@
 const { Users } = require("../dbObjects.js");
+const { customStatus } = require('../config.json');
 
 module.exports = {
   name: "ready",
@@ -6,6 +7,6 @@ module.exports = {
   async execute(client) {
     console.log(`Logged in as ${client.user.tag}!`);
 
-    client.user.setActivity("Repubblica Nassiryota", {type: 'WATCHING'});
+    client.user.setActivity(customStatus || "Repubblica Nassiryota", {type: 'WATCHING'});
   }
 };